pyproject.toml
[project]
name = "root-seeker"
version = "0.1.0"
description = "RootSeeker - AI-powered error analysis and root cause discovery (SLS + Zoekt + Qdrant + LLM)."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "httpx>=0.27",
  "pyyaml>=6.0",
  "aliyun-log-python-sdk>=0.9.30",
  "qdrant-client>=1.11",
  "tree-sitter>=0.25",
  "tree-sitter-python>=0.23",
  "tree-sitter-java>=0.23",
  "fastembed>=0.3",
]

[tool.setuptools.packages.find]
include = ["root_seeker*"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"
